The WebDisplay now supports overriding object colors and text from ANY ART module!  Use this to highlight LineSegments with lots of Clusters in them for example.  Fixed some bugs in the example and the data format itself which mans the version of UserHighlight got incremented at least once.  webDisplay_v2.html and WebDisplay_module.cc unfortunately got a bit more complicated.  I'll try to clean things up some other time.
This is the last feature I plan to add to "v2" of WebDisplay.  A few bug fix commits may follow as I test.

+++ EventDisplay/WebDisplay/UserHighlight.cxx
@@ -14,7 +14,7 @@ namespace evd
   {
   }
 
-  UserHighlight::UserHighlight(const short int hexColor, const std::string& comment): fHexColor(hexColor), fComment(comment)
+  UserHighlight::UserHighlight(const uint32_t hexColor, const std::string& comment): fHexColor(hexColor), fComment(comment)
   {
   }
 
@@ -23,7 +23,7 @@ namespace evd
     return fComment;
   }
 
-  const short int UserHighlight::Color() const
+  const uint32_t UserHighlight::Color() const
   {
     return fHexColor;
   }
@@ -33,7 +33,7 @@ namespace evd
     fComment = comment;
   }
 
-  void UserHighlight::SetColor(const short int color)
+  void UserHighlight::SetColor(const uint32_t color)
   {
     fHexColor = color;
   }
